requirements.txt
watchdog

ollama
langchain
langchain-ollama

pypdf2
python-docx